+++ src/tile_actor.h
@@ -12,7 +12,7 @@ public:
 
 	void SetTexture(sf::Texture* pTexture);
 	void SetTextureRect(const sf::IntRect& rect);
-	void SetPosition(float x, float y);
+	void SetPosition(int x, int y);
 
 	sf::IntRect GetRect() const;
